Get rid of a lint override
The issue is that passport returns a value of `any`, so the linter is
objecting to assigning that value of any to a variable (even if the
variable has a type). We can silence the lint by explicitly doing the
cast on the value, rather than implicitly casting when we perform the
assignment.