Split out various API contracts/clients
The primary goal here is limiting what we put into the client bundles.
In particular, we would like to keep the schemas and routes for the
Admin API and the (soon-to-be-created) Frontend API out of the
dependency graph reached by any of the client bundles, so we need to
split their definitions up a bit.