Add teamId to TeamState
This winds up being useful for the frontend to get the teamId from the
initial API response in the authRouter middleware.  In particular, this
allows passing the teamId to the frontend API client when marking gates
as satisfied in response to other requests for the Illegal Search round.
It's also needed for the websocket handler.