Elaborate on the WebSocket API
Adds client with reconnection logic and subscription sharing, and makes
use of the new pub/sub client/server ptotocol from the Stakeout round
page.

Eventually, we'll probably want to move some of this into a SharedWorker
so that multiple browser tabs will share the same WebSocket connection,
but something that does the baseline dynamic updates seems like a good
checkpoint.