Use a SharedWorker to mux multiple tabs pubsub across a single websocket
The webworker environment uses a different set of library types from the
one used by the dom environment, so we have to set up a separate
TypeScript project for those entrypoints.

Since the client code needs to reference the webworker script, we
declare it as a separate webpack target with its own entrypoint, so that
client code can look up that entrypoint from the manifest.  This makes
the dependency graph straightforward:

* worker stands alone
* client code depends on worker bundles through manifests
* server code depends on client bundles