Make the DatasetManager hot-reload its web worker
The websocket "hello" message from the server now includes the worker
script URL it expects the client to be using.  We propagate that through
so that the DatasetManager will spawn a new SharedWorker with the new
code, and re-establish subscriptions with that worker.

This helps ensure that we will not leak workers (and their connections)
even if we wind up deploying worker code changes during Hunt.

An unrelated cleanup is merging lib/websocket.ts ->
lib/api/websocket.ts; this was a mistake made in a previous commit where
I'd been flailing about getting the multiple-tsconfig setup going and
unintentionally duplicated the file.