Fix a few bugs and tweak semantics of SocketManager
The first bug/change is the most consequential: if you had multiple
subscribers to the same dataset, the deduplication of subscriptions sent
to the backend would cause only the first subscriber to receive a
callback with the initial subscription state.  While in theory we're
usually embedding initial states into the HTML, so this shouldn't
actually cause observably different behavior, the fact that the first
tab always gets the callback obscures that the embedded initial state is
load-bearing for other tabs.  Better to just be consistent.

As a result of fixing that bug, it becomes clear that reusing lastState
from a previous, now-stopped subscription when a new one is started is a
mistake, since it could frequently result in getting a first callback
for the watch() call with data *older* than even what we embedded in the
page, because the shared worker could have stopped the sub before the
current tab's initial HTTP request was served.  So I removed all the
logic around keeping the SubscriptionState object around once the sub is
stopped, and just removed the subs from the maps.

The second bug is that when SocketManager observes a socket close, it
would reset all subs to the "awaiting-socket" state, even if the sub was
in the stopped or stopping state.  As noted above, the stopped state is
no longer relevant, but the stopping state is still around, so if we
lost a websocket as we were trying to stop a subscription, then we'd
mark it as "awaiting-socket", which means that on reconnect we'd start
up a sub that had no watchers, which is wasteful.  We should only mark
subs that were "requesting" or "active" at the time we lost the socket
as waiting for socket readiness.

The third bug is that when a subscriber stops the subscription, we'd
always try to send the "unsub" packet over the websocket, even if the
socket was not alive.  We should do an RPC to cancel the sub if our
connection is live; otherwise we can simply mark the sub as stopped.