Make DevPane client-driven, and add a "dev" dataset
This allows DevPane to become live-updating and otherwise more useful
than the fully-SSR'd pane was.  By injecting the DevPane DOM nodes into
the document from client code at runtime, and changing how we position
the dev pane, we can also remove most of the Layout logic that changed
styles when rendering DevPane.

Now, the only difference between how we (server-side) render in dev mode
and how we (server-side) render for the actual site is that we include
one additional <script> tag when we want to include dev tools.

By using the websocket dataset publication scheme, we can also do a
couple other nice things: we can join in information from other sources
(like puzzle titles) which are not themselves included in TeamState, and
we can supply the dev pane with live updates, which means that now the
dev pane should now be accurate instead of stale.