Stakeout: switch to pointer events, with pointer capture
Pointer capture enables us to ensure we continue to receive move events
while dragging photos even if the drag leaves the window or the mouse
would pass over another element.  It's basically always what you want
when you're interactively dragging an element.

To enable us to use pointer capture effectively, rather than reordering
the `<div>`s for the photos when the stack order changes, we switch to
giving them z-indexes so that the elements do not change order, which
otherwise causes the pointer capture to break when React rerenders.