Add styled-components, swc plugin, and upgrade @swc/core
The `@swc/plugin-styled-components` plugin has to be version-compatible
with the `@swc/core`, so this involved bumping our `@swc/core` version
as well.

It appears that in a point version, the `@types/react` package removed
`Thenable` from being a top-level export of the `react` namespace, which
our vendored `types/mfng` files attempt to reference.  I removed them,
as they currently represent dead code and it seems unlikely we will use
mfng at this point.