Rework interaction graph structure
Add a type parameter for the expected keys of the speakers.  Because
Billie is never displayed in the scene (since we're seeing it through
Billie's eyes) , remove the `char_left` and `char_right` fields in favor
of a single `speaker` field that references the name and picture that
should be shown in the speech history area at the bottom.  To support
changing the art in the broader scene, we add the "overlay" global,
which should add a single layer with an alpha channel atop the
background.