Unify asset output path
This enables us to be sure to generate the exact same asset paths in
both SSR and client-rendered code, which means we can use hydrateRoot
instead of createRoot, which means we can allow browser React code to
reuse the SSR'd DOM instead of recreating it from scratch.

This is also strategically aligned with our goals around "have a single
static asset root from which we can serve assets", which the previous
solution with a variety of folders did not address.

One catch here is that because we have multiple configs writing to the
same output folder, we should explicitly avoid webpack's `clean`
option, which would throw away the outputs of earlier build steps.