Add full (cross-team) activity log API endpoint for frontends to tail/poll
This is an endpoint that should return the full relevant contents of the
activity log for all teams.  It is designed to support efficiently
maintaining a complete in-memory replica of the activity log dataset,
which will be useful for the websocket servers as they wish to push down
activity log updates which will drive transient in-site notifications.