Add dataset tailer which replicates activity log data
For now, we're focusing on the activity log dataset, but the
implementation is generic enough that we can use it with any
(append-only) dataset that has an `id` field which orders the rows,
which we can probably adapt the team_puzzle_guesses table to fit.

Some back-of-the-envelope math on how large we expect this dataset to
get during hunt:

  (number of teams * maximum activity log entries per team)
  + number of global actions (announcements, currency grants)

The latter basically rounds to zero.

In 2024, there were 345 teams registered, 91 of which were on campus.
266 teams solved at least one puzzle.
118 teams solved at least one underworld metapuzzle.

So, it's probably reasonable to assume that we will have fewer than 500
teams, and that of those, perhaps a third of teams will get far enough
to solve one metapuzzle.  So pessimistically assume that 1/3rd of those
teams will complete everything in the hunt (164 puzzle unlocks & solves,
4 interaction unlocks/starts/completes, 6 round unlocks, + however many
hint request & replies we can write) and we can probably conclude that
we're bounded above by:

166 * ((164 * 2) + (4 * 3) + (5)) = 57270 log entries from "strong" teams
 + 334 * (28 * 2) + 1) = 19038 log entries from "early-round" teams,
 excluding hint requests and replies.

At ~200 bytes per entry (probably less in practice), this is <15MiB for
the entire hunt.  This should be trivial to maintain in the websocket
servers.