Emit all activity log writes to pubsub
In general, we should not emit writes to pubsub until the transaction
containing them has committed, lest we create a situation where pubsub
readers will receive phantom writes.

To achieve this, instead of emitting writes to redis pubsub in the
middle of the database transaction, collect known insertions (for
log-shaped tables) or updates (for e.g. TeamState) that we want to emit
as we construct the transaction, and then once the transaction commits,
emit the objects to the appropriate pubsub channel as best-effort.  Even
if these writes are dropped, the dataset tailer will eventually discover
them through its periodic polling.