Make Shadow Diamond page also interactive
We don't really need the proof-of-concept CSS or image asset inclusion
any more; we have styled-components working in both SSR and client-side
rendered contexts, and image assets in both SSR and client-side
rendering are proven out in the Stakeout round, so we can make this a
little more useful now.

We can change the presentation later if we want, but this gives
something that connects the data end-to-end and allows unlocking puzzles
from the round page, rather than needing to use dev-mode-only pages.