Add frontend API for retrieving the full (cross-team) guess history
This dataset is expected to ultimately be similar in size to the
activity log.  The 2024 hunt site received something like 35000 guesses
over the course of the hunt; each row is expected to be something like
150-180 bytes of JSON, with (relatively rare, amongst the puzzles
dataset) canned responses potentially spiking another few hundred bytes.

All told, this dataset should also stay neatly under 16MiB for the
entire dataset for the entire hunt, so it's also a reasonable candidate
for replicating/tailing in the websocket server.