Publish activity log over pubsub
This involved a significant rework of the websocket server code to allow
for additional behavior and to make use of the DatasetTailer code from
previous commits.

Since all this only really works when we have redis, to simplify the
logic, the websocket handler is only defined when a `redisClient` is
provided.  Since the devtools are powered by a subscription that is
delivered over the websocket, this means displaying devtools also now
depends on redis.  I would recommend that we simply make redis (or an
equivalent pubsub backend) a hard requirement in the near future.