Subscriptions also take optional params
This will be needed for e.g. puzzle pages where we want to have
dataset="puzzle" but also need to pass which puzzle we want to monitor
the state of from client to the server.

For now, we define a single optional DatasetParams type, and backfill
`undefined` params everywhere we previously used watch().

We ensure that we consider both Dataset and DatasetParams when deduping
subscriptions; otherwise we'd wind up with all puzzle subscriptions
getting whatever the first puzzle sub's data was.