Move interaction definitions inside rounds
Eventually, we're probably going to want to link to the interaction from
the round page, so it would be convenient to have that data within the
round object we get in TeamState.  Additionally, this allows us to show
the interaction within the appropriate round in the devtools without
needing to duplicate that information in the devtools code, which seems
like a win.

It's marginally more work to check if a particular interaction slug is
unlocked, but that's still easy enough that it's fine.