Remove existing multiple-answer support
This approach never really had a good answer to the "how do we surface
submit constraints sanely?" problem, and at this point I'm willing to
give up on it.

The way we're currently envisioning presenting any puzzles with multiple
answers, Quentin's idea to simply have separate guess submission UI that
points at different slugs seems more viable.  We'll also be able to
reuse more of our other infrastructure by making separate submission
queues key to a separate slug -- pubsub for guess history will just do
the right thing, we can reuse frontend components that track the state
of the guess history, etc.

We *do* probably want to hide the slugs for the additional answer
puzzles from showing up by default in API responses or the activity log
or similar, lest they leak the notion at the API level as soon as the
gate that would unlock them is satisfied.

We'll likely need to add some sort of "related hidden puzzle" slug to
PuzzleDefinition (or perhaps PuzzleSlot, depending on where it makes the
most sense) to make it possible to find the related puzzle slug from the
puzzle that is considered to (nominally) have multiple answers.

While I was here, I also moved `prize` from the PuzzleDefinition
(alongside the `answer` to which it corresponds) back up to the slot.
The nature of wanting to award prizes for the first solve but not the
second led us to put `prize` by the answer, but it's really more often
the slot that determines whether or not we should award currency for
solves (mostly based on whether the slot is a meta or not).

As I did so, I discovered that we accepted guesses for puzzles which
were not included in the current hunt graph, so I made sure that we only
process guess submissions for valid slots (or development stubs).