Use generated identities for auto-increment primary key columns under PG
The Postgres-specific `serial` column type (which `table.increments`
uses under the hood on Postgres) is discouraged [1] in favor of using
generated identities.  SQLite supports generated columns, but not for
columns which are part of the primary key for the table [2], so we need
to pick schema behavior depending on the connection type.

[1] - https://wiki.postgresql.org/wiki/Don%27t_Do_This#Don.27t_use_serial
[2] - https://www.sqlite.org/gencol.html#limitations