Migrate build from webpack to rspack
Rspack [1] is a (mostly) drop-in reimplementation of webpack in Rust.
It is both more modern and more performant than webpack: production
builds complete on my machine with rspack in <1 second vs 4 seconds for
webpack, and even dev builds without nodeExternals complete at a similar
pace.

There were a couple places in the webpack config where rspack either
recommends doing things differently or did not support a config option
that webpack did.  This prompted me to do fewer output path shenanigans
and to just specify the output file as `static/<whatever>` for all the
assets we want to show up there, which allowed some simpler logic.

[1] - https://rspack.dev/