Implement canned responses (and rewards on canned responses)
We now allow canned responses to specify that a guess that matches a
particular canned response should provide the currency_delta (prize) for
that puzzle instead of the actual answer to the puzzle.  To support
this, we add an additional activity_log entry type of
`puzzle_partially_solved` so that we can attach the `currency_delta` to
that row.