Narrow props passed to PuzzleLink and use `<dialog>` for modal
Relying on the entire TeamState means that anything that renders
`PuzzleLink`s has to embed the whole TeamState in the initial
pagerender, which gets quite large as the hunt progresses.  Projecting
only the relevant fields for display purposes at the usage sites (and in
particular: only the data for the relevant puzzles displayed in that
context) will allow us to prepare smaller initial datasets.

`<dialog>` is supported by all recent browsers (Baseline 2022), and when
shown with `showModal()`, browsers do a very useful thing to manage
stacking order -- rendering the element on a top-layer, keeping the
element on top, and making a `::backdrop` pseudoelement, all of which
make making modal dialog box behavior much more convenient.

Concretely, this switch to `<dialog>` also happens to make for better
behavior of the modals on pages long enough to be scrollable (e.g. the
All Puzzles page, once we make that dynamic).