Synthesize an "id" field for the guess history pubsub dataset
For the `append`-type datasets, when we reconnect to the server after a
websocket disconnection, we re-establish the sub against the backend,
which results in us getting all the entries for the dataset again.

To avoid this reconnection resulting in appending every entry again, we
need some way to tell the difference between the entries we've already
received and the ones which are new.  One way to achieve that is to
require each entry to have some `id` field, and then to deduplicate on
`id`s.  Activity log entries already have ids, and strictly speaking, so
do guess log entries, but we don't expose them.  Synthesizing ids for
the guess history will let us use them for deduplication in SocketManager.