Add `title` to (most) activity log entries with `slug`
We need the user-facing title basically everywhere we'd want to use the
activity log contents, so there's little to gain by omitting it.  It's
possible that we could join in the titles only in the websocket server
codepath, so that the dataset tailers (which retain API responses) don't
waste memory holding a bunch of titles but that felt complicated for
relatively little gain.