Remove `answer_count` support from huntdata/logic
We've decided that we're going to implement "re-solve" puzzles as
separate guess queues, because otherwise it's hard to manage
rate-limits, block submitting the later solve until a prerequisite is
met, and also reject duplicate guess submissions.

Given that, `answer_count` is no longer a relevent piece of
functionality, so we can remove it to simplify the logic.