Be precise with internal and external activity log representations
We need the dataset tailers to handle the serialized-but-raw data schema
that we get from the DB and also shove into redis for pubsub.

We need the API and websocket to surface data which includes joining in
titles for puzzles, interactions, and gates.  The API also needs to
handle the fact that SQLite loads both dates and JSON as strings (and
the dates are missing the Z timezone format specifier).

Ultimately there are now basically four representations of activity_log
data:

1. the ActivityLogEntry format as described in the knex tables.  This is
   the most immediately usable format.
2. what we actually get back when we read from the DB or Redis (1, but
   with strings instead of JSON or Date objects)
3. the public API's getActivityLog entry type (timestamps as strings,
   `data` flattened, `title` joined in)
4. the internal API's getFullActivityLog entry type (timestamps as
   string, `internal_data` included as JSON, `data` included as JSON and
   not flattened into the entry object)

The easiest way to handle all of this is to have logic which uses format
1 internally, and then to convert the other formats to format 1 on load
and render from format 1 when serializing.  So that's what this does.