Round 1 super unlocks when all 4 interactions complete
We don't want to give folks the funaround until they have the four metas
that serve as inputs, but we also don't want to give people the "go meet
Billie!" at the exact same time as we give them the fourth dialogue tree
interaction.  So make sure the dialogue tree interactions are completed
first, and then we'll give solvers the new round and the funaround/super
together.