Enable skipLibCheck on worker tsconfig
There are some libraries (p5) that we want to use that assume they
always get typechecked in the `dom` environment, and have access to
things like `AudioNode` from the Web Audio API.  In the `webworker`
environment, those interfaces are not defined, so we hit errors if
we typecheck those libs as present in node_modules, even if we are
not referencing them or using them in any of the typechecked code.
These are not helpful to us, so we will turn them off.