Generate a content-addressable radio manifest
We want this manifest to appear in the statics folder and have a
content-addressed path, so we can have thingsboard just take notice when
the shared attribute changed to trigger prompt updates by all the radios.

We now also generate a top-level radio-manifest.json that points at the
(build-specific) filename.