Typeset tragic-spider (皇帝の暗号)
A fun exercise in slight dynamism without Javascript.

I transliterated the title from Kanji to ASCII for the puzzle slug.  I
suspect that nothing would actually break under the hood if we were to
use non-ASCII in the slug, but it's likely helpful for operator tools if
people can type some part of the slug without needing an IME, in case we
want to do typeahead search or something.