Typeset shoddy-table (The 10000-sheet Excel file)
Since certain file downloads want to preserve filenames, we create an
exceptions list in rspack.config.mjs of files for which we want to
preserve the final basename.  To preserve our ability to do aggressive
path-based caching, inside the static/ folder we create a directory
named by the file's hash, and emit the file with the original basename
into that directory.  So instead of

  dist/static/62072725f1ba212c44f3.xlsx

we will emit:

  dist/static/62072725f1ba212c44f3/10000sheets.xlsx