Clear all redis keys on startup
We were clearing both the global and per-team activity_log streams, but
not the sorted set that contains the high watermark values indicating
the max entry id that the per-team activity_log stream was expected to
contain.  This caused us to, after server shutdown and restart, to fail
to rewind the high watermark to the actual value, which caused us to
fail to retrieve entries from the stream, while also requesting that the
DB not return those entries either.