Suppress styled-component readable class names in prod builds
This has to be threaded through the styled-components SWC loader plugin,
which would normally read the SWC config from `.swcrc`.  We override the
configuration by reading the otherwise-default `.swcrc` into memory and
tweaking that config object.

Why not just put the whole config in rspack.config.mjs?  Because
`jest` also uses a jest-specific swc loader which also needs the config.
For unit tests, I don't think we care about dev vs prod mode (they'll
just get the default which is to provide the display names) but we do
want to be only tracking most of the SWC config in one place.