Typeset circular-greece (Educational Rite of Passage)
This involved making it possible to read TeamState from Puzzle or
Solution components, since we want to know if the team has picked
up the puzzle or not (which we model with a gate).

This also involved yet another importable resource format (this time,
pdf).  I'm checking in the models for the 3D printers; it's not yet
clear what we want to do with them (perhaps we will make a remote-only
version of the puzzle page that gives solvers the STLs?)