Link 3d printable parts
At some future point (after we run Hunt), we'll want to link these from
the puzzle page for the archives.  For now, let's link them from the
solution page, so we don't lose track of the content, and we can figure
the rest of that out later.

As we are adding more resource filetypes, we add more to the rspack
config, jest config, and asset-stub type definitions.