Make rspack compute fullhash by using [hash] in the asset/resource generator
In eb37e62c481f594f772090893c9ddd0f72567180 we regressed the radio
manifest generation by losing the `hash` field.

One of the performance optimizations that rspack performs is to avoid
computing hashes of resources if they are not going to be used by
plugins.  One of the ways it infers the use of the need of the hash is
by seeing if the filename string contains [hash] or [fullhash].

When I made the "asset/resource" generator read the field from
`pathData` and use it directly, I failed to realize that this meant that
I had unwittingly removed the way that rspack determines that the field
needs to be populated in the first place.

This is resolved by switching back to the [hash] substitution instead of
doing it myself, which means that rspack computes the full hash, and
then the RadioManifestPlugin will once again have that context available
to it on the objects exposed to the JS environment by the time it is
running and doing its final manifest generation, so we get the `hash`
field restored in the radio manifest (and the radio no longer errors out
because `hash` is a required field).