Paper Trail: use server-driven object states
There's some small regressions in behavior, which is why a few of the
DeskItems are still in Layout.tsx as a reminder to implement their
behavior.  But on the whole, this is basically the "wire it up to the
data and preserve confidentiality" patch.

I'm punting on the tooltip behavior for now.  Given the existing
"tooltips" getting covered up by the other objects, it seems clear to me
that we'll need a more comprehensive approach to make sure that we're
putting tooltips/popovers on a separate layer, and that the popovers for
unlockable puzzles are interactive.  And all that seems like it belongs
in a separate commit.