Make the murder round also client-side interactive
This round follows the same "set up a dataset, preload it, hydrate,
subscribe" flow that we've used to great effect elsewhere.

This isn't important yet since we have no puzzles for this round, but it
may become useful soon as more puzzles get written, and we may perhaps
typeset the meta soon.