Track "stray leads" puzzles
Experience would like to make sure that puzzles shown on the stray leads
("outlands") round page do not disappear when they are associated with
their investigation, but instead persist with a link to that round (once
the round becomes unlocked).

This means threading the Hunt data structure around a few more places,
since you have to have the Hunt to know what round a particular puzzle
is in, so you can know if it was unlocked at the time the puzzle became
unlocked.

I also fixed some wonkiness in the devtools which made testing this
manually not show what I expected in the devtools pane.