Avoid deleting teams that have already been registered on startup
Typesetters are generally running a workflow that looks like:

1. log in as the `unlocked` user
2. look at a puzzle page that you're in the process of typesetting
3. edit the puzzle component
4. restart the server
5. refresh the page

The change to team registration in seeds/demo.ts in
289fb965428d078ff7a74fdc6353d3e80850e327 was breaking this workflow: the
cookie issued in step 1 references a team_id that after step 4 no longer
existed in the DB, because the records for that team were being deleted,
but new insertions were generating new, larger ids.  This resulted in
the frontend erroring as the API returned 404 for /api/me -- the user
had a validly-signed cookie, but the cookie referenced a team_id the
backend no longer believed valid.  The worst part of this failure mode
is that it did not even allow self-recovery -- you'd have to manually
navigate to /logout to clear your cookie.

The purpose of the custom-state users is to enable typesetting workflows
and make it easier to validate server behavior.  If these users are
adding this much friction, they are not serving their purpose.

We probably shouldn't be wiping that much user state every server
startup.  Instead, we should make the edit-build-test loop tighter fix
up the specific things (the activity_log entries) that we think are
relevant to the use case for those custom-behavior users -- if a dev
wants to test registration or team info changes from a clean slate, they
can always `rm dev.sqlite3` or drop and recreate their DB.