ws: always dispatch initial teamState on new subscription
A browser that has been disconnected from the internet for some time and
is back online needs to catch up on updates via the websocket,
regardless of whether state changes are actively occurring.

We previously tried to elide initial updates on the websocket as an
optimization, but we assumed that the client state would closely follow
the state known to the server at the time of the websocket connection,
which for an initial pageload is true, but fails to hold in the
suspend-resume case.

We can always consider implementing an optimization that plumbs the
epoch all the way through from the client, and only dispatching on a
newer epoch, but for now, this will fix the incorrect observed behavior
and should still be pretty cheap.