Revise unlock structure per updated Puzzles team guidance
There's still a few places where we'll have to update feeder sets based
on which puzzles get assigned to which slots, but this is directionally
correct (visibility broadly based on solve counts within the round,
rather than unlock counts within a round).

We'll also have to adapt some things once we know which specific slots
will be released via the outlands.