Pubsub: gossip initial states and avoid backsliding
This change makes a series of small but noteworthy improvements to our
pubsub system.

First, we add an `epoch` field to every "replace" type dataset, and
ensure an `id` is present in every entry of "append" type datasets.
For datasets with epochs, we elide pushing epoch-only changes down the
websocket to avoid being unnecessarily chatty, just as we previously
elided pushing down other no-op changes.

Second, now that we have some way to compare freshness or identity of
dataset values, we only apply updates when the epoch exceeds our cached
epoch or (for append-type datasets) when we observe an entry with a
previously-unseen `id` value.  This prevents backsliding on pageload,
where the prerendered values on the page are newer than what is known to
the service worker.  (It's still slightly fraught to reset your database
or change logins without killing your shared worker, but the new
behavior is slightly better for most purposes.)

Third, we change the client interface to globalDatasetManager (and
related underlying functions) to also take the initial value of each
subscription, which is typically embedded in the page for initial
rendering.  Since we can now tell (from the epoch or id values) whether
the client has a newer or older value, we can update or merge any time a
new subscriber provides new initial data.  This enables us to propagate
updates across tabs even in the event that websockets are down if page
loads are still functional.

Lastly, we add a pair of hooks `useDataset` and `useAppendDataset` which
make consuming these auto-updating datasets a one-liner that triggers a
rerender on update, and switch to that at most callsites that used
globalDatasetManager.  A few callsites still want to do more complicated
things with callbacks, but most usages are radically simplified.