Increase redis isolated pool size above 1
Each DatasetTailer instance runs an executeIsolated() when setting up
watches on the streams we tail.  Each executeIsolated() consumes a
different isolated connection from the redis connection pool.  But the
default size of this connection pool is 1, so we were quietly failing to
observe any changes but those on the first dataset that we were trying
to tail.

Increasing the pool size means that the registration_team_log can also
be observed, and this means that the `team_info` subscriptions that were
previously failing to propagate updates now do so (and we'll also have
room for doing so with `puzzle_state_log` once I go implement that).

We also set `acquireTimeoutMillis` which will trigger errors if we try
to executeIsolated but have exhausted the isolated pool and cannot
acquire the resource.  This would have helped us find this failure mode
earlier instead of allowing it to just block indefinitely and quietly
fail to make progress.