Fix useAppendDataset
The bug is actually in DatasetManager, where we would inadvertently
change (in-place) the initialValue that the caller passed in.  Instead
we should do a shallow clone of the array, so that as we push updates to
the copy kept in DatasetManager, we do not modify our caller's object.

This had the effect of causing any caller of `useAppendDataset` to never
trigger a re-render when the data changed, because the `setState` call
observed the object already mutated out from under it to include the new
value, so it believed there was no change to apply.

Practically, this had the effect of causing the Activity Log page to not
update in realtime in response to new activity log entries, and would
also have caused similar issues with the puzzle state log as planned.