Export TeamRegistration{,State} from contract, not frontend_contract
The shapes are part of the public contract; we should not be importing
the types for their shapes from internal contract files.

Additionally, we need to ensure the API methods for retrieving a team
registration return an epoch alongside the registration state.
Otherwise, it will be impossible for pubsub datasets to be comparable to
API responses, which means we'd be unable to make API requests for
initial page renders that we would be able to hydrate and use with
`useDataset`.