ws server handles projections of registration log
While team_info is content with just `name`, the `/team` page is going
to want more of the current registration data.  This patch restructures
subscriptions that want to tail the team registration log to accept a
callback mapper function on the reduced TeamInfoIntermediate.

We then define a new pubsub dataset, `team_registration`, which contains
the full (latest) team registration data.