Fix performance issue in use{,Append}Dataset with params
If you pass a params object, each call to useDataset (or
useAppendDataset) would likely have been passed a different object
instance, so React (using `Object.is()`) would have considered the
params to have changed.

What we actually care about is if the params changed semantically, which
we can achieve by stringifying the params and seeing if those differ.
(We could also do a deep object compare, but this is easier and lets us
piggyback on the existing deps array behavior.)

This change fixes a significant performance issue observed in the
not-yet-merged new-ketchup (What Do They Call You?) implementation,
where every single message would trigger a rerender, which would trigger
a cancellation and recreation of the subscription.  Because the effect's
behavior was still correct (just unperformant), the page would
eventually load, but this way it will stream the whole history in rather
than incurring an additional roundtrip or two of latency for each
message in the log, which added up quickly.