Implement satisfied_if: for gates in huntdata
We want the blacklight gate to be automatically satisfied when a team
solves the second Illegal Search meta.  This change adds a way to define
conditions to automatically mark a gate as satisfied, updates the logic
to compute when that should be the case, and updates the data layer to
write back such satisfied gates when another activity log write would
satisfy the conditions.