Tweak Bookcase styles for readability
Reducing line-height and using slightly larger fonts for the titles
should make this slightly more legible.  I kept the author size at 12px
and added a small amount of margin between the two lines to help make
the distinction between the book title and the author visually distinct.

There was one book title that overflowed to three lines at 14px, so I
make it possible to override the display size and set that one to 12px.