Rework bookcase view
The original conception of the bookcase dated from when we assumed we
were going to be going for quasi-photorealistic across the Illegal
Search. However, given that the rest of our interactable artwork is in a
more cartoonish style, this reworks the bookcase to be a little more in
line with that.

Instead of looking at the books straight on, we get a little bit of a
perspective, so we can see the tops and covers of the books. When they
are pulled in or out, they slide on an isometric plane instead of just
getting scaled up and down.