Add support for copy-to-clipboard
This adds a new property to puzzle content definition, `copyable`. If
set to true, the server-side render injects a `data-copyable` attribute
onto the puzzle content element. Client side code then looks for the
presence of this attribute and if present adds a copy-to-clipboard
button.

(I also used this as an opportunity to separate the content type for
puzzles vs. solutions - it's hard to imagine that we're going to want to
inject either copy-paste support or client-side Javascript into
solutions.)

The button itself does some arcane magic to reformat the puzzle content
in a way that Google Sheets and Excel are more likely to accept before
stashing it on the clipboard. It does this by creating a temporary clone
of the content with a few specific changes:

- Anything with class no-copy is removed, and anything with class
  copy-only (hidden by default) is shown. This allows individual puzzles
  and components to override the markup that gets copied when necessary.
- Any ordered lists are (somewhat sloppily) replaced with tables as
  otherwise the list numbering is lost.
- A fixed whitelist of styles are copied over from the original content.
- Any time an element has non-default styling, we insert a "reset span"
  after it (otherwise Google Sheets will sometimes carry over the
  styling to the next cell).
- We attempt to render images (and links to images) as in-cell embeds,
  although this only works in very specific circumstances. Outside of
  those circumstances, we fall back to links.

For a handful of puzzles I've been using for testing, this is sufficient
to make copy-to-clipboard work well with no additional changes, so
enable for those.