Skip typechecking eslint rules on puzzles and solutions
The actual puzzle and solution files are just rendering (at least
relatively) static content, but cost us an awful lot of time to lint, so
skip the type-aware lint checkers.

This is a compromise to preserve the generally-useful type-aware
linting, but skip it on files that are unlikely to have the sorts of
bugs that the type-aware linting would catch. (Especially given that
we're visually inspecting the output of the code)